refactor: rm header component
