--- src/app/post/[pubky]/[postId]/components/_Header.tsx
@@ -1,22 +1,0 @@
-'use client';
-
-import * as Components from '@/components';
-import { Icon } from '@social/ui-shared';
-import { useRouter } from 'next/navigation';
-
-export function Header() {
-  const router = useRouter();
-
-  return (
-    <>
-      <Components.Header title="Post" />
-      <Components.HeaderMobile
-        leftIcon={
-          <div className="cursor-pointer" onClick={() => router.back()}>
-            <Icon.ArrowLeft size="20" />
-          </div>
-        }
-      />
-    </>
-  );
-}
